refactor(viewer): extract viewer_competition.jsx from viewer.jsx
Pure file split, no behavior change. Step 9 of mp-pxxc.
