+++ web-mobile/index.html
@@ -104,6 +104,10 @@
          buildWatchlistUpcoming), extracted from viewer.jsx (mp-pxxc step 7).
          Must load after viewer_match.js and viewer_watchlist_core.js, and before viewer.js. -->
     <script type="module" src="/dist/viewer_schedule.js?v=5"></script>
+    <!-- Competition detail + overview page components (ViewerCompetition, ViewerOverview),
+         extracted from viewer.jsx (mp-pxxc step 9).
+         Must load after viewer_schedule.js and before viewer.js. -->
+    <script type="module" src="/dist/viewer_competition.js?v=5"></script>
     <script type="module" src="/dist/viewer.js?v=5"></script>
     <script type="module" src="/dist/admin_helpers.js?v=5"></script>
     <script type="module" src="/dist/admin_shell.js?v=5"></script>
